Be more conservative in snapshot and rgl.pixels calculations.
git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/rgl/pkg/rgl@1436 79a1b648-8b68-254f-b20b-c0b50f5137bf

+++ src/Device.h
@@ -34,7 +34,7 @@ public: // -- all methods are blocking until action completed
   bool open(void); // -- if failed, instance is invalid and should be deleted
   void close(void); // -- when done, instance is invalid and should be deleted
   bool snapshot(int format, const char* filename);
-  bool pixels(int* ll, int* size, int component, float* result);
+  bool pixels(int* ll, int* size, int component, double* result);
   bool postscript(int format, const char* filename, bool drawText);
 
   bool clear(TypeID stackTypeID);
